make layout module
