+++ pyhdtoolkit/plotting/lattice.py
@@ -11,15 +11,16 @@ from typing import Optional, Tuple
 import matplotlib
 import matplotlib.axes
 import matplotlib.pyplot as plt
+import pandas as pd
 
 from cpymad.madx import Madx
 from loguru import logger
 
+from pyhdtoolkit.plotting.layout import plot_machine_layout
 from pyhdtoolkit.plotting.utils import (
     _get_twiss_table_with_offsets_and_limits,
     make_survey_groups,
     maybe_get_ax,
-    plot_machine_layout,
 )
